custom-elements/registering-custom-elements/unresolved-elements-interface-test-002.html

<!DOCTYPE html>
<html>
<head>
<title>Unresolved element interface must be SVGElement, if the namespace is SVG Namespace</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="author" title="Vasiliy Degtyarev" href="mailto:vasya@unipro.ru">
<meta name="assert" content="When an unresolved element is created, it's element interface must be SVGElement if the namespace is SVG Namespace">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#registering-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    var xsvg = doc.createElementNS(SVG_NAMESPACE, 'x-svg');
    
    assert_class_string(xsvg, 'SVGElement', 'Element must be a SVG element');

}, 'Unresolved element interface with valid name for SVG namespace');


test(function() {
    var doc = newHTMLDocument();	
    doc.body.innerHTML = '<svg xmlns=' + SVG_NAMESPACE + ' version="1.1">' +
    				'<x-svg-1 id="xsvg"></x-svg-1>' +
    				'</svg>';
    var xsvg = doc.querySelector('#xsvg');
    				
    assert_class_string(xsvg, 'SVGElement', 'Element must be a SVG element');
    
}, 'Unresolved element interface with valid name is instantiating in HTML for SVG namespace');
</script>
</body>
</html>
